+++ src/main/java/ru/autosome/commons/backgroundModel/di/DiBackground.java
@@ -106,7 +106,7 @@ public class DiBackground implements DiBackgroundModel {
   }
 
   @Override
-  public double volume() {
+  public int volume() {
     return 1;
   }
